chore: fix storybooks
