+++ botfront/.storybook/config.js
@@ -1,13 +1,19 @@
 import { configure, addDecorator } from '@storybook/react';
 import 'semantic-ui-css/semantic.min.css';
 import '../client/main.less';
-import { withProvider } from './decorators';
+import {
+    withReduxProvider,
+    withProjectContext,
+    withStoriesContext,
+} from './decorators';
 
 function loadStories() {
     const req = require.context('../stories', true, /\.stories\.js$/);
     req.keys().forEach(filename => req(filename));
 }
 
-addDecorator(withProvider);
+addDecorator(withReduxProvider);
+addDecorator(withProjectContext);
+addDecorator(withStoriesContext);
 
 configure(loadStories, module);
