+++ botfront/.storybook/decorators.js
@@ -1,8 +1,154 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Provider } from 'react-redux';
+import { ProjectContext } from '../imports/ui/layouts/context';
+import { ConversationOptionsContext } from '../imports/ui/components/stories/Context';
 import store from '../imports/ui/store/store';
+import { OOS_LABEL } from '../imports/ui/components/constants.json';
 
-export const withProvider = (story) => (
+export const slots = [
+    { name: 'textSlot1', type: 'text' }, { name: 'textSlot2', type: 'text' }, { name: 'textSlot3', type: 'text' },
+    { name: 'catSlot1', type: 'categorical', categories: ['c1'] }, { name: 'catSlot2', type: 'categorical', categories: ['c2'] }, { name: 'catSlot3', type: 'categorical', categories: ['c3'] },
+    { name: 'listSlot1', type: 'list' }, { name: 'listSlot2', type: 'list' }, { name: 'listSlot3', type: 'list' },
+    { name: 'boolSlot1', type: 'bool' }, { name: 'boolSlot2', type: 'bool' }, { name: 'boolSlot3', type: 'bool' },
+];
+
+const responseOne = {
+    key: 'utter_yay',
+    values: [{
+        lang: 'en',
+        sequence: [
+            { content: 'text: YAY!!' },
+            { content: 'text: BOO!!' },
+            { content: 'text: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.' },
+        ],
+    }],
+};
+
+const responseTwo = {
+    key: 'utter_boo',
+    values: [{
+        lang: 'en',
+        sequence: [
+            { content: 'text: I love peanutes too' },
+            { content: 'text: Can I call you l8r' },
+            { content: 'text: <3' },
+        ],
+    }],
+};
+
+const intentsFixture = ['intent1', 'intent2', 'intent3', 'intent4'];
+const entitiesFixture = ['entity1', 'entity2', 'entity3', 'entity4'];
+
+const utteranceOne = {
+    intent: 'I_come_from',
+    entities: [
+        {
+            start: 12,
+            end: 17,
+            value: 'Paris',
+            entity: 'from',
+        },
+    ],
+    text: `I come from Paris. Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.`,
+};
+
+const utteranceTwo = {
+    intent: 'I_want_peanuts_on_my_sundae',
+    entities: [
+        {
+            start: 7,
+            end: 14,
+            value: 'peanuts',
+            entity: 'food',
+        },
+        {
+            start: 43,
+            end: 50,
+            value: 'peanuts',
+            entity: 'food',
+        },
+    ],
+    text: 'I love peanuts all night long. Do you love peanuts as much as me? Ok ?',
+};
+
+const utteranceThree = {
+    intent: 'Im_summarizing_Moby_Dick',
+    entities: [],
+    text: `On the first day of the chase, Ahab smells the whale, climbs the mast, and sights Moby Dick. He claims the doubloon for himself, and orders all
+boats to lower except for Starbuck's. The whale bites Ahab's boat in two, tosses the captain out of it, and scatters the crew. On the second day of the chase,
+Ahab leaves Starbuck in charge of the Pequod. Moby Dick smashes the three boats that seek him into splinters and tangles their lines. Ahab is rescued, but his ivory leg and Fedallah are lost.
+Starbuck begs Ahab to desist, but Ahab vows to slay the white whale, even if he would have to dive through the globe itself to get his revenge.`,
+};
+
+const responseFixtures = [responseOne, responseTwo];
+const utterances = { utteranceOne, utteranceTwo, utteranceThree };
+
+const parseUtterance = u => ({
+    intent: OOS_LABEL,
+    text: u,
+});
+
+const getResponse = key => ({
+    key,
+    values: [{
+        lang: 'en',
+        sequence: [
+            { content: 'text: fetched response!!' },
+            { content: 'text: fetched response!!' },
+            { content: 'text: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.' },
+        ],
+    }],
+});
+
+const getUtteranceFromPayload = (intent, callback) => {
+    callback(
+        null,
+        [
+            ...Object.values(utterances).filter(u => u.intent === intent.intent),
+            { intent, entities: [], text: 'not found' },
+        ][0],
+    );
+};
+
+export const withProjectContext = (story) => {
+    const [responses, updateResponses] = useState(responseFixtures);
+    const [entities, setEntities] = useState(entitiesFixture);
+    const [intents, setIntents] = useState(intentsFixture);
+
+    const addEntity = entity => setEntities(Array.from(new Set([...entities, entity])));
+    const addIntent = intent => setIntents(Array.from(new Set([...intents, intent])));
+
+    return (
+        <ProjectContext.Provider
+            value={{
+                slots,
+                intents,
+                entities,
+                responses,
+                parseUtterance,
+                addEntity,
+                addIntent,
+                getResponse,
+                getUtteranceFromPayload,
+                updateResponses,
+                language: 'en',
+            }}
+        >
+            {story()}
+        </ProjectContext.Provider>
+    );
+};
+
+export const withStoriesContext = (story) => (
+    <ConversationOptionsContext.Provider
+        value={{ storyGroups: [] }}
+    >
+        {story()}
+    </ConversationOptionsContext.Provider>
+);
+
+export const withReduxProvider = (story) => (
     <Provider store={store}>
         {story()}
     </Provider>
