+++ botfront/.storybook/knobs.js
@@ -1,0 +1,27 @@
+export const size = {
+    mini: 'mini',
+    tiny: 'tiny',
+    small: 'small',
+    medium: 'medium',
+    large: 'large',
+    big: 'big',
+    huge: 'huge',
+    massive: 'massive',
+};
+
+export const color = {
+    red: 'red',
+    orange: 'orange',
+    yellow: 'yellow',
+    olive: 'olive',
+    green: 'green',
+    teal: 'teal',
+    blue: 'blue',
+    violet: 'violet',
+    purple: 'purple',
+    pink: 'pink',
+    brown: 'brown',
+    grey: 'grey',
+    black: 'black',
+    botfrontBlue: 'botfront-blue',
+};
