+++ botfront/stories/UserUtterancePopupContent.stories.js
@@ -1,10 +1,8 @@
+/* eslint-disable no-alert */
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { withKnobs } from '@storybook/addon-knobs';
-import { ProjectContext } from '../imports/ui/layouts/context';
 import UserUtterancePopupContent from '../imports/ui/components/stories/common/UserUtterancePopupContent';
 import DashedButton from '../imports/ui/components/stories/common/DashedButton';
-import { intents, entities } from './AddStoryLine.stories';
 
 const alertPayload = pl => alert(`
     Intent: ${pl.intent}
@@ -16,14 +14,6 @@ const alertPayload = pl => alert(`
 const trigger = <DashedButton color='blue'>User says:</DashedButton>;
 
 storiesOf('UserUtterancePopupContent', module)
-    .addDecorator(withKnobs)
-    .addDecorator(story => (
-        <div className='story-visual-editor'>
-            <ProjectContext.Provider value={{ intents, entities }}>
-                {story()}
-            </ProjectContext.Provider>
-        </div>
-    ))
     .add('default', () => (
         <UserUtterancePopupContent
             onCreateFromInput={() => alert('from input!!')}
